Grouped all functions accessing ensemble.
