+++ libenkf/src/enkf_main.c
@@ -187,6 +187,8 @@ void enkf_main_update_local_updates( enkf_main_type * enkf_main);
 static void enkf_main_close_fs( enkf_main_type * enkf_main );
 static void enkf_main_init_fs( enkf_main_type * enkf_main );
 static void enkf_main_user_select_initial_fs(enkf_main_type * enkf_main );
+static void enkf_main_free_ensemble( enkf_main_type * enkf_main );
+static void enkf_main_init_jobname( enkf_main_type * enkf_main);
 
 /*****************************************************************/
 
@@ -209,11 +211,6 @@ bool enkf_main_set_refcase( enkf_main_type * enkf_main , const char * refcase_pa
 
 ui_return_type * enkf_main_validata_refcase( const enkf_main_type * enkf_main , const char * refcase_path) {
   return ecl_config_validate_refcase( enkf_main->ecl_config , refcase_path );
-}
-
-void enkf_main_init_jobname( enkf_main_type * enkf_main) {
-  for (int iens = 0; iens < enkf_main->ens_size; iens++)
-    enkf_state_update_jobname( enkf_main->ensemble[iens] );
 }
 
 
@@ -352,16 +349,6 @@ ui_return_type * enkf_main_set_data_file( enkf_main_type * enkf_main , const cha
 
 
 
-static void enkf_main_free_ensemble( enkf_main_type * enkf_main ) {
-  if (enkf_main->ensemble != NULL) {
-    const int ens_size = enkf_main->ens_size;
-    int i;
-    for (i=0; i < ens_size; i++)
-      enkf_state_free( enkf_main->ensemble[i] );
-    free(enkf_main->ensemble);
-    enkf_main->ensemble = NULL;
-  }
-}
 
 
 void enkf_main_free(enkf_main_type * enkf_main){
@@ -457,15 +444,6 @@ static vector_type * enkf_main_alloc_node_ensemble(const enkf_main_type * enkf_m
 
 
 
-
-enkf_state_type * enkf_main_iget_state(const enkf_main_type * enkf_main , int iens) {
-  return enkf_main->ensemble[iens];
-}
-
-
-member_config_type * enkf_main_iget_member_config(const enkf_main_type * enkf_main , int iens) {
-  return enkf_state_get_member_config( enkf_main->ensemble[iens] );
-}
 
 
 
@@ -2269,64 +2247,6 @@ runpath_list_type * enkf_main_get_runpath_list( const enkf_main_type * enkf_main
 }
 
 
-/**
-   This function will resize the enkf_main->ensemble vector,
-   allocating or freeing enkf_state instances as needed.
-*/
-
-
-void enkf_main_resize_ensemble( enkf_main_type * enkf_main , int new_ens_size ) {
-  int iens;
-
-  /* No change */
-  if (new_ens_size == enkf_main->ens_size)
-    return ;
-
-  ranking_table_set_ens_size( enkf_main->ranking_table , new_ens_size );
-  /* Tell the site_config object (i.e. the queue drivers) about the new ensemble size: */
-  site_config_set_ens_size( enkf_main->site_config , new_ens_size );
-
-
-  /* The ensemble is shrinking. */
-  if (new_ens_size < enkf_main->ens_size) {
-    /*1: Free all ensemble members which go out of scope. */
-    for (iens = new_ens_size; iens < enkf_main->ens_size; iens++)
-      enkf_state_free( enkf_main->ensemble[iens] );
-
-    /*2: Shrink the ensemble pointer. */
-    enkf_main->ensemble = util_realloc(enkf_main->ensemble , new_ens_size * sizeof * enkf_main->ensemble );
-    enkf_main->ens_size = new_ens_size;
-    return;
-  }
-
-
-  /* The ensemble is expanding */
-  if (new_ens_size > enkf_main->ens_size) {
-    /*1: Grow the ensemble pointer. */
-    enkf_main->ensemble = util_realloc(enkf_main->ensemble , new_ens_size * sizeof * enkf_main->ensemble );
-
-    /*2: Allocate the new ensemble members. */
-    for (iens = enkf_main->ens_size; iens < new_ens_size; iens++)
-
-      /* Observe that due to the initialization of the rng - this function is currently NOT thread safe. */
-      enkf_main->ensemble[iens] = enkf_state_alloc(iens,
-                                                   enkf_main->rng ,
-                                                   model_config_iget_casename( enkf_main->model_config , iens ) ,
-                                                   enkf_main->pre_clear_runpath                                 ,
-                                                   int_vector_safe_iget( enkf_main->keep_runpath , iens)        ,
-                                                   enkf_main->model_config                                      ,
-                                                   enkf_main->ensemble_config                                   ,
-                                                   enkf_main->site_config                                       ,
-                                                   enkf_main->ecl_config                                        ,
-                                                   enkf_main->templates                                         ,
-                                                   enkf_main->subst_list);
-    enkf_main->ens_size = new_ens_size;
-    return;
-  }
-
-  util_abort("%s: something is seriously broken - should NOT be here .. \n",__func__);
-}
-
 
 void enkf_main_add_node(enkf_main_type * enkf_main, enkf_config_node_type * enkf_config_node) {
     for (int iens = 0; iens < enkf_main_get_ensemble_size(enkf_main); iens++) {
@@ -2826,15 +2746,6 @@ int enkf_main_get_ensemble_size( const enkf_main_type * enkf_main ) {
 }
 
 
-enkf_state_type ** enkf_main_get_ensemble( enkf_main_type * enkf_main) {
-  return enkf_main->ensemble;
-}
-
-
-const enkf_state_type ** enkf_main_get_ensemble_const( const enkf_main_type * enkf_main) {
-  return (const enkf_state_type **) enkf_main->ensemble;
-}
-
 
 
 /**
@@ -3219,5 +3130,5 @@ void enkf_main_export_ranking(enkf_main_type * enkf_main, const char * ranking_k
   ranking_table_fwrite_ranking(ranking_table, ranking_key, ranking_file);
 }
 
-
+#include "enkf_main_ensemble.c"
 #include "enkf_main_manage_fs.c"
